Reuse cycle counter between experiments
