+++ jitter/exp1.cpp
@@ -1,21 +1,9 @@
 #include <iostream>
 #include <vector>
 #include <cstdint>
-
+#include "exp.hpp"
 
 using namespace std;
-
-__attribute__((noinline))
-void task(int iterations) {
-  volatile int i = 0;
-  while(i < iterations) i++;
-}
-
-inline uint64_t cycles() noexcept {
-  uint32_t hi, lo;
-  asm("rdtscp" : "=a"(lo), "=d"(hi));
-  return ((uint64_t) lo) | ((uint64_t) hi) << 32;
-}
 
 std::vector<uint64_t> cycle_counter;
